+++ mocks/github-client.go
@@ -29,6 +29,20 @@ type GitHubClient struct {
 		result1 github.PRBranches
 		result2 error
 	}
+	PRDiffStub        func(context.Context, string) (string, error)
+	pRDiffMutex       sync.RWMutex
+	pRDiffArgsForCall []struct {
+		arg1 context.Context
+		arg2 string
+	}
+	pRDiffReturns struct {
+		result1 string
+		result2 error
+	}
+	pRDiffReturnsOnCall map[int]struct {
+		result1 string
+		result2 error
+	}
 	PRStateStub        func(context.Context, string) (string, string, string, error)
 	pRStateMutex       sync.RWMutex
 	pRStateArgsForCall []struct {
@@ -145,6 +159,71 @@ func (fake *GitHubClient) GetPRBranchesReturnsOnCall(i int, result1 github.PRBra
 	}
 	fake.getPRBranchesReturnsOnCall[i] = struct {
 		result1 github.PRBranches
+		result2 error
+	}{result1, result2}
+}
+
+func (fake *GitHubClient) PRDiff(arg1 context.Context, arg2 string) (string, error) {
+	fake.pRDiffMutex.Lock()
+	ret, specificReturn := fake.pRDiffReturnsOnCall[len(fake.pRDiffArgsForCall)]
+	fake.pRDiffArgsForCall = append(fake.pRDiffArgsForCall, struct {
+		arg1 context.Context
+		arg2 string
+	}{arg1, arg2})
+	stub := fake.PRDiffStub
+	fakeReturns := fake.pRDiffReturns
+	fake.recordInvocation("PRDiff", []interface{}{arg1, arg2})
+	fake.pRDiffMutex.Unlock()
+	if stub != nil {
+		return stub(arg1, arg2)
+	}
+	if specificReturn {
+		return ret.result1, ret.result2
+	}
+	return fakeReturns.result1, fakeReturns.result2
+}
+
+func (fake *GitHubClient) PRDiffCallCount() int {
+	fake.pRDiffMutex.RLock()
+	defer fake.pRDiffMutex.RUnlock()
+	return len(fake.pRDiffArgsForCall)
+}
+
+func (fake *GitHubClient) PRDiffCalls(stub func(context.Context, string) (string, error)) {
+	fake.pRDiffMutex.Lock()
+	defer fake.pRDiffMutex.Unlock()
+	fake.PRDiffStub = stub
+}
+
+func (fake *GitHubClient) PRDiffArgsForCall(i int) (context.Context, string) {
+	fake.pRDiffMutex.RLock()
+	defer fake.pRDiffMutex.RUnlock()
+	argsForCall := fake.pRDiffArgsForCall[i]
+	return argsForCall.arg1, argsForCall.arg2
+}
+
+func (fake *GitHubClient) PRDiffReturns(result1 string, result2 error) {
+	fake.pRDiffMutex.Lock()
+	defer fake.pRDiffMutex.Unlock()
+	fake.PRDiffStub = nil
+	fake.pRDiffReturns = struct {
+		result1 string
+		result2 error
+	}{result1, result2}
+}
+
+func (fake *GitHubClient) PRDiffReturnsOnCall(i int, result1 string, result2 error) {
+	fake.pRDiffMutex.Lock()
+	defer fake.pRDiffMutex.Unlock()
+	fake.PRDiffStub = nil
+	if fake.pRDiffReturnsOnCall == nil {
+		fake.pRDiffReturnsOnCall = make(map[int]struct {
+			result1 string
+			result2 error
+		})
+	}
+	fake.pRDiffReturnsOnCall[i] = struct {
+		result1 string
 		result2 error
 	}{result1, result2}
 }